refactor: split dlt_common and dlt_log
Seperate all internal logging function into an own source file.
This reduces the scope of the sprawling dlt_common source file.
