+++ Android.bp
@@ -99,6 +99,7 @@ cc_binary {
         "src/gateway/dlt_gateway.c",
         "src/lib/dlt_client.c",
         "src/shared/dlt_common.c",
+        "src/shared/dlt_log.c",
         "src/shared/dlt_config_file_parser.c",
         "src/shared/dlt_multiple_files.c",
         "src/shared/dlt_offline_trace.c",
@@ -142,6 +143,7 @@ cc_library_shared {
         "src/lib/dlt_user.c",
         "src/shared/dlt_common.c",
         "src/shared/dlt_multiple_files.c",
+        "src/shared/dlt_log.c",
         "src/shared/dlt_protocol.c",
         "src/shared/dlt_user_shared.c",
     ],
