+++ src/console/dlt-receive.c
@@ -83,6 +83,7 @@
 #   include <limits.h>
 #endif
 #include <inttypes.h>
+#include "dlt_log.h"
 #include "dlt_client.h"
 #include "dlt-control-common.h"
