+++ src/daemon/dlt_daemon_event_handler.c
@@ -34,6 +34,7 @@
 #include <syslog.h>
 
 #include "dlt_common.h"
+#include "dlt_log.h"
 
 #include "dlt-daemon.h"
 #include "dlt-daemon_cfg.h"
