+++ src/offlinelogstorage/dlt_offline_logstorage_behavior.c
@@ -29,7 +29,7 @@
 #include <errno.h>
 #include <libgen.h>
 
-#include "dlt_common.h"
+#include "dlt_log.h"
 #include "dlt_offline_logstorage.h"
 #include "dlt_offline_logstorage_behavior.h"
 #include "dlt_offline_logstorage_behavior_internal.h"
