+++ tests/gtest_dlt_daemon_multiple_files_logging.cpp
@@ -30,6 +30,7 @@ int connectServer(void);
 
 extern "C"
 {
+#include "dlt_log.h"
 #include "dlt_common.h"
 #include <syslog.h>
 #include <dirent.h>
